alteraçoes

+++ lavajato_estoque/models.py
@@ -7,8 +7,8 @@ class produto(models.Model):
     nome = models.CharField(max_length=200)
     valor_venda = models.DecimalField(max_digits=10, decimal_places=2)
     valor_compra = models.DecimalField(max_digits=10, decimal_places=2)
-    quantidade = models.IntegerField()
-    quantidade_minima = models.IntegerField()
+    quantidade = models.DecimalField(max_digits=10, decimal_places=2)
+    quantidade_minima = models.DecimalField(max_digits=10, decimal_places=2)
     lucro = models.IntegerField(null=True, blank=True)
     data_cadastro = models.DateTimeField(default=timezone.now)
     
@@ -20,7 +20,7 @@ class lote(models.Model):
     prod = models.ForeignKey(produto,on_delete=models.CASCADE)
     valor_venda = models.DecimalField(max_digits=10, decimal_places=2)
     valor_compra = models.DecimalField(max_digits=10, decimal_places=2)
-    quantidade = models.IntegerField()
+    quantidade = models.DecimalField(max_digits=10, decimal_places=2)
     data_cadastro = models.DateTimeField(default=timezone.now)
     
     def __str__(self):
@@ -34,7 +34,7 @@ class retiradas(models.Model):
     id = models.AutoField(primary_key=True)
     uso = models.CharField(max_length=1, choices=OP)
     prod = models.ForeignKey(produto,on_delete=models.CASCADE)
-    quantidade = models.IntegerField()
+    quantidade = models.DecimalField(max_digits=10, decimal_places=2)
     data = models.DateTimeField(default=timezone.now)
     
     def __str__(self):
